Move and rename all the things
