+++ lib/view_component/storybook/stories_v1.rb
@@ -1,0 +1,150 @@
+# frozen_string_literal: true
+
+module ViewComponent
+  module Storybook
+    class StoriesV1
+      extend ActiveSupport::DescendantsTracker
+      include ActiveModel::Validations
+
+      class_attribute :story_configs, default: []
+      class_attribute :stories_parameters, :stories_title, :stories_layout
+
+      validate :validate_story_configs
+
+      class << self
+        def title(title = nil)
+          # if no argument is passed act like a getter
+          self.stories_title = title unless title.nil?
+          stories_title
+        end
+
+        def story(name, component = default_component, &block)
+          story_config = StoryConfig.new(story_id(name), name, component, layout, &block)
+          story_config.instance_eval(&block)
+          story_configs << story_config
+          story_config
+        end
+
+        def parameters(params = nil)
+          # if no argument is passed act like a getter
+          self.stories_parameters = params unless params.nil?
+          stories_parameters
+        end
+
+        def layout(layout = nil)
+          # if no argument is passed act like a getter
+          self.stories_layout = layout unless layout.nil?
+          stories_layout
+        end
+
+        def to_csf_params
+          validate!
+          csf_params = { title: title }
+          csf_params[:parameters] = parameters if parameters.present?
+          csf_params[:stories] = story_configs.map(&:to_csf_params)
+          csf_params
+        end
+
+        def write_csf_json
+          json_path = File.join(stories_path, "#{stories_name}.stories.json")
+          File.write(json_path, JSON.pretty_generate(to_csf_params))
+          json_path
+        end
+
+        def stories_name
+          name.chomp("Stories").underscore
+        end
+
+        # Returns all component stories classes.
+        def all
+          load_stories if descendants.empty?
+          descendants
+        end
+
+        # Returns +true+ if the stories exist.
+        def stories_exists?(stories_name)
+          all.any? { |stories| stories.stories_name == stories_name }
+        end
+
+        # Find a component stories by its underscored class name.
+        def find_story_configs(stories_name)
+          all.find { |stories| stories.stories_name == stories_name }
+        end
+
+        # Returns +true+ if the story of the component stories exists.
+        def story_exists?(name)
+          story_configs.map(&:name).include?(name.to_sym)
+        end
+
+        # find the story by name
+        def find_story_config(name)
+          story_configs.find { |config| config.name == name.to_sym }
+        end
+
+        # validation - ActiveModel::Validations like but on the class vs the instance
+        def valid?
+          # use an instance so we can enjoy the benefits of ActiveModel::Validations
+          @validation_instance = new
+          @validation_instance.valid?
+        end
+
+        delegate :errors, to: :@validation_instance
+
+        def validate!
+          valid? || raise(ValidationError, @validation_instance)
+        end
+
+        private
+
+        def inherited(other)
+          super(other)
+          # setup class defaults
+          other.stories_title = Storybook.stories_title_generator.call(other)
+          other.story_configs = []
+        end
+
+        def default_component
+          name.chomp("Stories").constantize
+        end
+
+        def load_stories
+          Dir["#{stories_path}/**/*_stories.rb"].sort.each { |file| require_dependency file } if stories_path
+        end
+
+        def stories_path
+          Storybook.stories_path
+        end
+
+        def story_id(name)
+          "#{stories_name}/#{name.to_s.parameterize}".underscore
+        end
+      end
+
+      protected
+
+      def validate_story_configs
+        story_configs.reject(&:valid?).each do |story_config|
+          story_errors = story_config.errors.full_messages.join(', ')
+          errors.add(:story_configs, :invalid_story, story_name: story_config.name, story_errors: story_errors)
+        end
+
+        story_names = story_configs.map(&:name)
+        duplicate_names = story_names.group_by(&:itself).map { |k, v| k if v.length > 1 }.compact
+        return if duplicate_names.empty?
+
+        duplicate_name_sentence = duplicate_names.map { |name| "'#{name}'" }.to_sentence
+        errors.add(:story_configs, :duplicate_stories, count: duplicate_names.count, duplicate_names: duplicate_name_sentence)
+      end
+
+      class ValidationError < StandardError
+        attr_reader :stories
+
+        def initialize(stories)
+          @stories = stories
+
+          super("#{@stories.class.name} invalid: (#{@stories.errors.full_messages.join(', ')})")
+        end
+      end
+    end
+  end
+end
