+++ lib/view_component/storybook/story.rb
@@ -3,15 +3,30 @@
 module ViewComponent
   module Storybook
     class Story
-      include ActiveModel::Validations
 
-      attr_reader :component, :content_block, :slots, :layout
+      attr_reader :id, :name, :parameters, :controls
 
-      def initialize(component, content_block, slots, layout)
-        @component = component
-        @content_block = content_block
-        @slots = slots
-        @layout = layout
+      def initialize(id, name, parameters, controls)
+        @id = id
+        @name = name
+        @parameters = parameters
+        @controls = controls
+      end
+
+      def to_csf_params
+        csf_params = { name: name, parameters: { server: { id: id } } }
+        csf_params.deep_merge!(parameters: parameters) if parameters.present?
+        controls.each do |control|
+          csf_params.deep_merge!(control.to_csf_params)
+        end
+        csf_params
+      end
+
+      def self.from_code_object(code_object, stories_config)
+        name = code_object.name
+        id = "#{stories_config.stories_name}/#{name.to_s.parameterize}".underscore
+
+        self.new(id, name, {}, [])
       end
     end
   end
