+++ spec/dummy/test/components/stories/content_component_stories.rb
@@ -1,33 +1,40 @@
 # frozen_string_literal: true
 
 class ContentComponentStories < ViewComponent::Storybook::Stories
-  story :with_string_content do
-    content "Hello World!"
-  end
 
-  story :with_control_content do
-    content text("Hello World!")
-  end
-
-  story :with_described_control do
-    content text("Hello World!").description('My first computer program.')
-  end
-
-  story :with_block_content do
-    content do
+  def with_string_content
+    render(ContentComponent.new) do
       "Hello World!"
     end
   end
 
-  story :with_helper_content do
-    content do
-      link_to "Hello World!", "#"
+  # controls(only: :with_described_control) do
+  #   text(:content, "Hello World")
+  # end
+
+  control :content, as: :text, default: "Hello World!", only: :with_control_content
+
+  def with_control_content(content: "Hello World!")
+    render(ContentComponent.new) do
+      content
     end
   end
 
-  story :with_constructor_content do
-    constructor do
-      "Hello World!"
+  # controls(only: :with_described_control) do
+  #   text(:content, "Hello World", description: "My first computer program.")
+  # end
+
+  control :content, as: :text, default: "Hello World!", description: "My first computer program.", only: :with_described_control
+
+  def with_described_control(content: "Hello World!")
+    render(ContentComponent.new) do
+      content
+    end
+  end
+
+  def with_helper_content
+    render(ContentComponent.new) do
+      content_tag(:span, "Hello World!")
     end
   end
 end
