+++ spec/dummy/test/components/stories/demo/heading_component_stories.rb
@@ -4,10 +4,14 @@ module Demo
   class HeadingComponentStories < ViewComponent::Storybook::Stories
     title 'Heading Component'
 
-    story :default do
-      constructor(
-        heading_text: text("Heading")
-      )
+    # controls do
+    #   text(:heading_text, "Heading")
+    # end
+
+    control :heading_text, as: :text, default: "Heading"
+
+    def default(heading_text: "Heading")
+      render HeadingComponent.new(heading_text: button_text)
     end
   end
 end
