+++ spec/dummy/test/components/stories/dry_component_stories.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 class DryComponentStories < ViewComponent::Storybook::Stories
-  story :default do
-    constructor(
-      title: text("Hello World!")
-    )
+  control :title, as: :text, default: "Hello World!"
+
+  def default(title: "Hello World!")
+    render DryComponent.new(title: title)
   end
 end
