--- spec/dummy/test/components/stories/legacy_controls_dsl_stories.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-class LegacyControlsDslStories < ViewComponent::Storybook::Stories
-  story :short_button, Demo::ButtonComponent do
-    controls do
-      button_text "OK"
-    end
-  end
-end
