+++ spec/dummy/test/components/stories/no_layout_stories.rb
@@ -3,16 +3,14 @@
 class NoLayoutStories < ViewComponent::Storybook::Stories
   layout false
 
-  story :default, Demo::ButtonComponent do
-    constructor(
-      button_text: text("OK")
-    )
+  # @control button_text text
+  def default(button_text: "OK")
+    render Demo::ButtonComponent.new(button_text: button_text)
   end
 
-  story :mobile_layout, Demo::ButtonComponent do
-    constructor(
-      button_text: text("OK")
-    )
-    layout "mobile"
-  end
+  # # @control button_text text
+  # # @layout mobile
+  # def mobile_layout(button_text: "OK")
+  #   render Demo::ButtonComponent.new(button_text: button_text)
+  # end
 end
